requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
docxtpl==0.20.0
python-docx==1.2.0
lxml==4.9.3
pandas==2.2.3
matplotlib==3.8.2
openpyxl==3.1.2
docx2pdf==0.1.8
